[MOVE/#189] 아워투두 리스트 폴더 수정
